[egs,scripts] Fix bugs in Dihard 2018 (#2897)

+++ egs/wsj/s5/steps/nnet3/train_raw_dnn.py
@@ -321,7 +321,7 @@ def train(args, run_opts):
     num_archives_expanded = num_archives * args.frames_per_eg
     num_archives_to_process = int(args.num_epochs * num_archives_expanded)
     num_archives_processed = 0
-    num_iters = ((num_archives_to_process * 2)
+    num_iters = int((num_archives_to_process * 2)
                  / (args.num_jobs_initial + args.num_jobs_final))
 
     # If do_final_combination is True, compute the set of models_to_combine.
